Copy and paste some CAF stuff, make one edit, rinse / repeat
Oh wait, I forgot that isn't what actually happens, except when adding
120 commits to take a camera that has limited functionality and convert
it (key word) into something compatible with a camera that should never
have skipped those commits to begin with. Actually, we are adding
credits to kushdeck for his work on the HAL (props), mirroring the HAL
to 8660 and beginning the process to make it work. This is based on the
information provided that the HP Touchpad and Evo3d are almost hardware
identical internally. By extension, the base code should be roughly
similar to allow minor modification for the secondary camera. This was
not done by CAF, so I guess it is far beyond my capability, but then
again neither was the 7x30 HAL we use, but somehow that still manages
to work ;)